Fix bug in context with type classes and floating values
All float and int values in a program are stored in the value environment with a name starting with "_". Due to a missing pattern in `renameLiteral` ( src/Checks/SyntaxCheck) the initial name "_" was never extended with a random identifier. That is, all float values were named "_" and could not be identified as temporary values and, thus, were exported in the interface file. A function named "_" in the interface files caused a runtime error ("Unexpected token").